warnings cleaned up

+++ _studio/shared/umc/core/umc/include/umc_defs.h
@@ -637,7 +637,7 @@ VideoDecoder *CreateAsynchronousVideoDecoder(VideoDecoder *pDecoder);
 #endif //__cplusplus
 
 #include "ippdefs.h"
-#define MFX_INTERNAL_CPY(dst, src, size) ippsCopy_8u((const Ipp8u *)(src), (Ipp8u *)(dst), size)
+#define MFX_INTERNAL_CPY(dst, src, size) ippsCopy_8u((const Ipp8u *)(src), (Ipp8u *)(dst), (int)size)
 
 /******************************************************************************/
